disciplinas/04-criacao-de-agentes/projects/007-de-mock-para-real/runtime/requirements.txt
openai>=1.0
pyyaml
python-dotenv
requests
fastapi
uvicorn
